backend/requirements.txt
fastapi
uvicorn[standard]
cassandra-driver
pydantic-settings
